+++ xill-processor/src/main/java/nl/xillio/xill/plugins/date/services/DateService.java
@@ -151,4 +151,22 @@ public interface DateService extends XillService {
      * @return The date corresponding with the timestamp
      */
     Date fromTimestamp(long timestamp);
+
+    /**
+     * Determines if the date is before another date
+     *
+     * @param date1 First comparison date
+     * @param date2 Second comparison date
+     * @return True if the date1 is before date2, otherwise false
+     */
+    boolean isBefore(Date date1, Date date2);
+
+    /**
+     * Determines if the date is after another date
+     *
+     * @param date1 First comparison date
+     * @param date2 Second comparison date
+     * @return True if the date1 is after date2, otherwise false
+     */
+    boolean isAfter(Date date1, Date date2);
 }